requirements.txt
google-generativeai>=0.3.0
selenium>=4.16.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.0
requests>=2.31.0
flask>=3.0.0 